内联编辑器基本完成
https://github.com/codebdy/rxdrag/issues/214
