+++ examples/inline-editor-example/src/page/locales.ts
@@ -1,0 +1,23 @@
+export const pageLocales = {
+  "zh-CN": {
+    title: "页面",
+    settings: {
+    }
+
+  },
+  'en-US': {
+    title: "Page",
+    settings: {
+    }
+  }
+}
+
+
+export const pageResourceLocales = {
+  "zh-CN": {
+    "Page": "页面",
+  },
+  'en-US': {
+    "Page": "Page",
+  }
+}
