+++ examples/inline-editor-example/src/page/schema.ts
@@ -1,0 +1,4 @@
+import { createSchema } from "@rxdrag/react-antd-materials";
+import { INodeSchema } from "@rxdrag/schema";
+
+export const paeSchema: INodeSchema = createSchema()
