+++ packages/rxdrag/editor/react-antd-shell-inline/src/components/OutlinePanel/index.tsx
@@ -5,6 +5,7 @@ import { Panel } from "../Panel"
 import { WidgetNames } from "../../interfaces"
 import { useClose } from "../../hooks/useClose"
 import { useSettersTranslate } from "@rxdrag/react-core"
+import { OutlineTree } from "@rxdrag/react-antd-shell"
 
 const Container = styled(Panel).attrs({ name: WidgetNames.outline })`
   left:${8 + 260}px;
@@ -19,6 +20,7 @@ export const OutlinePanel = memo(() => {
       <PanelTitle onClose={close}>
         {t("outline")}
       </PanelTitle>
+      <OutlineTree />
     </Container>
   )
 })
