+++ packages/rxdrag/editor/react-antd-shell/src/widgets/index.ts
@@ -12,3 +12,4 @@ export * from "./SelectionButtions"
 export * from "./ThemeButton"
 export * from "./UndoRedoButtons"
 export * from "./ViewButtons"
+export * from "./OutlineWidget/OutlineTree"
